# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API MAK3RS DOCUMENTATION",
        "description": "",
        "contact": {
            "email": "info@mak3rs.tk"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/communities/all": {
            "get": {
                "tags": [
                    "Community"
                ],
                "description": "Obtenemos todas las comunidades",
                "operationId": "App\\Http\\Controllers\\CommunityController::communities",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "alias": {
                                        "description": "",
                                        "type": "string"
                                    },
                                    "name": {
                                        "description": "",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "List Communities"
                    }
                }
            }
        },
...
        "/communities/create": {
            "post": {
                "tags": [
                    "Community"
                ],
                "description": "Creamos la comunidad",
                "operationId": "App\\Http\\Controllers\\CommunityController::create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "alias": {
                                        "description": "",
                                        "type": "string"
                                    },
                                    "name": {
                                        "description": "",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Object Community or null"
                    }
                }
            }
        },
        "/communities/update": {
            "put": {
                "tags": [
                    "Community"
                ],
                "description": "Actualizamos la comunidad",
                "operationId": "App\\Http\\Controllers\\CommunityController::update",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "uuid": {
                                        "description": "",
                                        "type": "string"
                                    },
                                    "name": {
                                        "description": "",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Object Community or null"
                    }
                }
            }
        },
        "/communities/delete": {
            "delete": {
                "tags": [
                    "Community"
                ],
                "description": "Borramos la comunidad",
                "operationId": "App\\Http\\Controllers\\CommunityController::delete",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "uuid": {
                                        "description": "",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                }
            }
        },
...
        "/users/communities": {
            "get": {
                "tags": [
                    "User"
                ],
                "description": "Listado de comunidades a las que pertenece el usuario",
                "operationId": "App\\Http\\Controllers\\UserController::communities",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "uuid": {
                                        "description": "User",
                                        "type": "string"
                                    },
                                    "alias": {
                                        "description": "",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        }
    }
}